smart_plant_care/requirements.txt
# 核心深度学习框架
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0

# 强化学习库
gymnasium>=0.29.0  # OpenAI Gym的现代替代品
stable-baselines3>=2.0.0  # PPO等RL算法的高质量实现
sb3-contrib>=2.0.0  # 额外的RL算法

# 数据可视化
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
plotly>=5.14.0

# 科学计算
scipy>=1.10.0
pandas>=2.0.0

# 配置管理
pyyaml>=6.0
hydra-core>=1.3.0

# 代码质量
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

# Jupyter支持
jupyter>=1.0.0
ipykernel>=6.23.0
ipywidgets>=8.0.0

# 进度条
tqdm>=4.65.0

# 视频录制（用于演示）
opencv-python>=4.7.0
moviepy>=1.0.3

# 命令行界面
click>=8.1.0
rich>=13.3.0

# GPU监控（可选）
nvidia-ml-py3>=7.352.0

# 性能分析
psutil>=5.9.0

